+++ src/main/java/com/gestionFinanzas/Categories/Investment/InvestmentCategoryRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface InvestmentCategoryRepository extends JpaRepository<InvestmentCategory, Integer> {
+public interface InvestmentCategoryRepository extends JpaRepository<InvestmentCategory, Long> {
