+++ src/main/java/com/gestionFinanzas/Troncales/IncomeOrExpense/RecurrenceDetails/RecurrenceDetailsRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface RecurrenceDetailsRepository extends JpaRepository<RecurrenceDetails, Integer> {
+public interface RecurrenceDetailsRepository extends JpaRepository<RecurrenceDetails, Long> {
 
     @Query(value = "SELECT r FROM RecurrenceDetails r WHERE (r.endDate IS NULL OR r.endDate >= CURRENT_DATE)")
     List<RecurrenceDetails> findActiveRecurrences();
